Add simple bookings "database"
The database consists of a json-file protected behind an RwLock. Will
almost certainly need to be changed into a real database at some point
but this is good for now.

The database is generic over the stored type so can be used for storing
more/other things in the future.