Extract TelescopeController from SalsaTelescope
We want to achieve two things:
- Shrinking SalsaTelescope which is getting big
- More intuitive updating of telescope controller state and tracking

To do this we take out the parts relating to the telescope controller
from SalsaTelescope. The SalsaTelescope remains as a facade but can in
the future be broken up further.

By using a long lived task to do all communication to with one telescope
controller we prepare for using less large locks in the future. Updating
the telescope continuously (e.g while tracking) is moved into this task.
This allows other parts of the system to send commands and observe the
effects on the telescope without having to continuously drive it.