Extract TelescopeController from SalsaTelescope (#71)
We want to achieve two things:
- Shrinking SalsaTelescope which is getting big
- More intuitive updating of telescope controller state and tracking

To do this we take out the parts relating to the telescope controller
from SalsaTelescope. The SalsaTelescope remains and keeps control of
the receiver for now.

By using a long lived task to do all communication to with one telescope
controller we prepare for using less large locks in the future. Updating
the telescope continuously (e.g while tracking) is moved into this task.
This allows other parts of the system to send commands and observe the
effects on the telescope without having to continuously drive it.