Implement error handling in observe
Put a div on the page and put any errors into the div.

It will not quite work if no page with an error box has ever been
rendered. This will require some further reworking.

Part of the error handling here could be extracted to deal with errors
that occur elsewhere.