Set telescope to observe through from path (#123)
Use `observe/<telescope id>` to select which telescope to use for
observation. This required some random updates to not break other
things.

The _Observe_ link in the navbar is now pointed to `observe/fake` to
keep the same behavior. Other telescopes are only available through
changing the path directly in the browser.