Fix measurement websocket to wrong telescope (#142)
The measurement websocket URL is extracted from the current
window.location instead of hardcoding to `fake`. This makes it possible
to observe a spectrum with a real telescope.

Also change `fake` to two fake telescopes: `fake1` and `fake2`. Having
two different ones is helpful for testing out telescope selection logic.