Put models and routes in own folders (#177)
Let's try to separate the more pure logic that goes in models from the
messy reality of HTTP and HTML that is going on in routes.

Ideally models can depend on models but not routes. Routes can depend on
models but ideally *not* on other routes. This is not completely there
yet.

There is a further need to have somewhere to put general routes-support code.

Telescope control is it's own thing and shouldn't go into the models
module. We can consider this to be on a lower abstraction level. Same
goes for general database code etc.